# WIP- DO NOT MERGE: Make head Postings only return series in time range