--- a/tsdb/compact.go
+++ b/tsdb/compact.go
@@ -682,7 +682,7 @@ func (c *LeveledCompactor) populateBlock(blocks []BlockReader, meta *BlockMeta, 
 			}
 		}
 
-		indexr, err := b.Index()
+		indexr, err := b.Index(math.MinInt64, math.MaxInt64)
 		if err != nil {
 			return errors.Wrapf(err, "open index reader for block %s", b)
 		}