--- a/tsdb/head_bench_test.go
+++ b/tsdb/head_bench_test.go
@@ -14,6 +14,7 @@
 package tsdb
 
 import (
+	"fmt"
 	"strconv"
 	"sync/atomic"
 	"testing"
@@ -47,4 +48,36 @@ func BenchmarkHeadStripeSeriesCreateParallel(b *testing.B) {
 			h.getOrCreate(uint64(i), labels.FromStrings("a", strconv.Itoa(int(i))))
 		}
 	})
+}
+
+func BenchmarkHeadSeries(b *testing.B) {
+	h, err := NewHead(nil, nil, nil, 1000)
+	testutil.Ok(b, err)
+	defer h.Close()
+	app := h.Appender()
+	numSeries := 1000000
+	for i := 0; i < numSeries; i++ {
+		app.Add(labels.FromStrings("foo", "bar", "i", strconv.Itoa(i)), int64(i), 0)
+	}
+	testutil.Ok(b, app.Commit())
+
+	matcher := labels.MustNewMatcher(labels.MatchEqual, "foo", "bar")
+
+	for s := 1; s <= numSeries; s *= 10 {
+		b.Run(fmt.Sprintf("%dof%d", s, numSeries), func(b *testing.B) {
+			q, err := NewBlockQuerier(h, 0, int64(s-1))
+			testutil.Ok(b, err)
+
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				ss, err := q.Select(matcher)
+				testutil.Ok(b, err)
+				for ss.Next() {
+				}
+				testutil.Ok(b, ss.Err())
+			}
+			q.Close()
+		})
+
+	}
 }